#66 interceptor에서 comment, post 업로드시 임시 저장 심어야됨

+++ src/main/java/kr/seoul/snsX/service/PostServiceImpl.java
@@ -90,11 +90,15 @@ public class PostServiceImpl implements PostService {
 
     @Override
     @Transactional
-    public void removePost(Long pk) throws EntityNotFoundException, FileNotFoundException {
+    public void removePost(Long pk, Long memberId) throws EntityNotFoundException, FileNotFoundException {
         Post foundPost = postRepository.findById(pk)
                 .orElseThrow(() -> new EntityNotFoundException());
-        fileRepository.deleteFiles(foundPost.getImages());
-        postRepository.delete(foundPost);
+        if (havePermission(memberId, foundPost.getMember().getId())) {
+            fileRepository.deleteFiles(foundPost.getImages());
+            postRepository.delete(foundPost);
+        } else {
+            throw new InvalidException();
+        }
     }
 
     @Override
@@ -116,10 +120,17 @@ public class PostServiceImpl implements PostService {
                 .orElseThrow(() -> new EntityNotFoundException());
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new EntityNotFoundException());
-        if (memberId == comment.getMember().getId())
+        if (havePermission(memberId, comment.getMember().getId()))
             commentRepository.delete(comment);
         else
             throw new InvalidException();
+    }
+
+    private boolean havePermission(Long memberId, Long makerId) {
+        if (memberId == makerId)
+            return true;
+        else
+            return false;
     }
 
     @Override
